requirements.txt
sentence-transformers
faiss-cpu
langchain_text_splitters
numpy
python-dotenv
requests